backend/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "This project is a quick-start example for building a web scraping API using FastAPI and Hatchet for task management."
authors = []
readme = "README.md"

[tool.poetry.scripts]
start-api = "src.api.main:start"
start-worker = "src.workflows.main:start"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.0"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
hatchet-sdk = "^0.35.1"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
